+++ Domain/RDD.Domain.Tests/PropertySelectorEqualityComparerTests.cs
@@ -22,7 +22,7 @@ namespace RDD.Domain.Tests
         [Fact]
         public void ComparingTheSameReferencedPropertyShouldWork()
         {
-            var dictionary = new Dictionary<IExpression, int>(new RddIExpressionEqualityComparer());
+            var dictionary = new Dictionary<IExpression, int>(new RddExpressionEqualityComparer());
             var p1 = ExpressionChain<FakeClass>.New(fc => fc.A);
             var p1Val = 42;
 
@@ -34,7 +34,7 @@ namespace RDD.Domain.Tests
         [Fact]
         public void ComparingDifferentPropertyInstancesShouldWork()
         {
-            var dictionary = new Dictionary<IExpression, int>(new RddIExpressionEqualityComparer());
+            var dictionary = new Dictionary<IExpression, int>(new RddExpressionEqualityComparer());
             var p1 = ExpressionChain<FakeClass>.New(fc => fc.A);
             var p1Val = 42;
             dictionary.Add(p1, p1Val);
@@ -47,7 +47,7 @@ namespace RDD.Domain.Tests
         [Fact]
         public void ComparingPropertyInstancesWithDifferentEntityNameButSamePropertyShouldWork()
         {
-            var dictionary = new Dictionary<IExpression, int>(new RddIExpressionEqualityComparer());
+            var dictionary = new Dictionary<IExpression, int>(new RddExpressionEqualityComparer());
             var p1 = ExpressionChain<FakeClass>.New(fc => fc.A);
             var p1Val = 42;
             dictionary.Add(p1, p1Val);
@@ -60,7 +60,7 @@ namespace RDD.Domain.Tests
         [Fact]
         public void ComparingTwoDifferentPropertiesShouldFail()
         {
-            var dictionary = new Dictionary<IExpression, int>(new RddIExpressionEqualityComparer());
+            var dictionary = new Dictionary<IExpression, int>(new RddExpressionEqualityComparer());
             var p1 = ExpressionChain<FakeClass>.New(fc => fc.A);
             var p1Val = 42;
             dictionary.Add(p1, p1Val);
@@ -73,7 +73,7 @@ namespace RDD.Domain.Tests
         [Fact]
         public void ComparingPropertiesWithTheSameNameFromTwoDifferentClassesShouldFail()
         {
-            var dictionary = new Dictionary<IExpression, int>(new RddIExpressionEqualityComparer());
+            var dictionary = new Dictionary<IExpression, int>(new RddExpressionEqualityComparer());
             var p1 = ExpressionChain<FakeClass>.New(fc => fc.A);
             var p1Val = 42;
             dictionary.Add(p1, p1Val);