+++ Domain/RDD.Domain/Exceptions/TechnicalException.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Net;
+using System.Runtime.Serialization;
 
 namespace RDD.Domain.Exceptions
 {
@@ -7,7 +8,8 @@ namespace RDD.Domain.Exceptions
     /// Should be used for internal / technical exceptions only. 
     /// Will be translated as HttpStatusCode.InternalServerError : 500
     /// </summary>
-    public class TechnicalException : Exception, IStatusCodeException
+    [Serializable]
+    public sealed class TechnicalException : ApplicationException, IStatusCodeException
     {
         public TechnicalException(string message)
             : base(message)
@@ -19,6 +21,17 @@ namespace RDD.Domain.Exceptions
         {
         }
 
-        public virtual HttpStatusCode StatusCode => HttpStatusCode.InternalServerError;
+        private TechnicalException(SerializationInfo info, StreamingContext context)
+            : base(info, context)
+        {
+        }
+
+        public HttpStatusCode StatusCode => HttpStatusCode.InternalServerError;
+
+        public override void GetObjectData(SerializationInfo info, StreamingContext context)
+        {
+            base.GetObjectData(info, context);
+            info.AddValue(nameof(StatusCode), StatusCode);
+        }
     }
 }