+++ Domain/RDD.Domain/Exceptions/UnauthorizedException.cs
@@ -1,15 +1,19 @@
 ﻿using System;
 using System.Net;
+using System.Runtime.Serialization;
 
 namespace RDD.Domain.Exceptions
 {
-    public class UnauthorizedException : BusinessException
+    [Serializable]
+    public sealed class UnauthorizedException : BusinessException
     {
-        public UnauthorizedException(string message) : base(message)
+        public UnauthorizedException(string message)
+            : base(message)
         {
         }
 
-        public UnauthorizedException(string message, Exception innerException) : base(message, innerException)
+        private UnauthorizedException(SerializationInfo info, StreamingContext context)
+            : base(info, context)
         {
         }