+++ Domain/RDD.Domain/Exceptions/UnreachableEntityException.cs
@@ -1,11 +1,18 @@
 ﻿using System;
+using System.Runtime.Serialization;
 
 namespace RDD.Domain.Exceptions
-{ 
-    public class UnreachableEntityException : BadRequestException
+{
+    [Serializable]
+    public sealed class UnreachableEntityException : BadRequestException
     {
-        public UnreachableEntityException(Type entityType) 
+        public UnreachableEntityException(Type entityType)
             : base($"Unreachable entity type {entityType.Name}. Consider adding Combinations to your Application.")
+        {
+        }
+
+        private UnreachableEntityException(SerializationInfo info, StreamingContext context)
+            : base(info, context)
         {
         }
     }