+++ Domain/RDD.Domain/Exceptions/UnsupportedContentTypeException.cs
@@ -1,17 +1,19 @@
 ﻿using System;
 using System.Net;
+using System.Runtime.Serialization;
 
 namespace RDD.Domain.Exceptions
 {
-    public class UnsupportedContentTypeException : BusinessException
+    [Serializable]
+    public sealed class UnsupportedContentTypeException : BusinessException
     {
-        public UnsupportedContentTypeException(string message) 
+        public UnsupportedContentTypeException(string message)
             : base(message)
         {
         }
 
-        public UnsupportedContentTypeException(string message, Exception innerException) 
-            : base(message, innerException)
+        private UnsupportedContentTypeException(SerializationInfo info, StreamingContext context)
+            : base(info, context)
         {
         }