+++ Domain/RDD.Domain/Patchers/DynamicPatcher.cs
@@ -6,22 +6,13 @@ namespace RDD.Domain.Patchers
 {
     class DynamicPatcher : IPatcher
 	{
-        object IPatcher.InitialValue(PropertyInfo property, object patchedObject)
-        {
-            return null;
-        }
+        object IPatcher.InitialValue(PropertyInfo property, object patchedObject) => null;
 
-		object IPatcher.PatchValue(object patchedObject, Type expectedType, IJsonElement json)
+        object IPatcher.PatchValue(object patchedObject, Type expectedType, IJsonElement json)
 		{
 			return PatchValue(patchedObject, expectedType, json as JsonObject);
 		}
 
-		public virtual object PatchValue(object patchedObject, Type expectedType, JsonObject json)
-		{
-			if (json == null)
-				return null;
-
-			return json.GetContent();
-		}
-	}
+        public virtual object PatchValue(object patchedObject, Type expectedType, JsonObject json) => json?.GetContent();
+    }
 }