+++ Domain/RDD.Domain/Rights/ICombinationsHolder.cs
@@ -1,7 +1,7 @@
 ﻿using RDD.Domain.Models;
 using System.Collections.Generic;
 
-namespace RDD.Domain
+namespace RDD.Domain.Rights
 {
     public interface ICombinationsHolder
     {